Extension method testing
